// ts/src/ledger/network_pb.ts
// @generated by protoc-gen-es v2.5.1 with parameter "target=ts"
// @generated from file ledger/network.proto (package api.ledger, syntax proto3)
/* eslint-disable */

import type { GenEnum, GenFile } from "@bufbuild/protobuf/codegenv2";
import { enumDesc, fileDesc } from "@bufbuild/protobuf/codegenv2";

/**
 * Describes the file ledger/network.proto.
 */
export const file_ledger_network: GenFile = /*@__PURE__*/
  fileDesc("ChRsZWRnZXIvbmV0d29yay5wcm90bxIKYXBpLmxlZGdlciq6AQoHTmV0d29yaxIVChFVTkRFRklORURfTkVUV09SSxAAEhMKD1NURUxMQVJfTkVUV09SSxADEhMKD0JJVENPSU5fTkVUV09SSxAFEhQKEExJVEVDT0lOX05FVFdPUksQBxIUChBFVEhFUkVVTV9ORVRXT1JLEAkSEgoOUklQUExFX05FVFdPUksQCxIcChhTQV9TVE9DS19CUk9LRVJTX05FVFdPUksQDxIQCgxOVUxMX05FVFdPUksQEEIkWiJnaXRodWIuY29tL21lc2h0cmFkZS9hcGkvZ28vbGVkZ2VyYgZwcm90bzM");

/**
 *
 * Network represents the various blockchain networks supported by Mesh.
 * NOTE: numbering NOT TO BE UPDATED to be sequential without an associated migration!!!!
 *
 * @generated from enum api.ledger.Network
 */
export enum Network {
  /**
   * @generated from enum value: UNDEFINED_NETWORK = 0;
   */
  UNDEFINED_NETWORK = 0,

  /**
   * @generated from enum value: STELLAR_NETWORK = 3;
   */
  STELLAR_NETWORK = 3,

  /**
   * @generated from enum value: BITCOIN_NETWORK = 5;
   */
  BITCOIN_NETWORK = 5,

  /**
   * @generated from enum value: LITECOIN_NETWORK = 7;
   */
  LITECOIN_NETWORK = 7,

  /**
   * @generated from enum value: ETHEREUM_NETWORK = 9;
   */
  ETHEREUM_NETWORK = 9,

  /**
   * @generated from enum value: RIPPLE_NETWORK = 11;
   */
  RIPPLE_NETWORK = 11,

  /**
   * @generated from enum value: SA_STOCK_BROKERS_NETWORK = 15;
   */
  SA_STOCK_BROKERS_NETWORK = 15,

  /**
   * @generated from enum value: NULL_NETWORK = 16;
   */
  NULL_NETWORK = 16,
}

/**
 * Describes the enum api.ledger.Network.
 */
export const NetworkSchema: GenEnum<Network> = /*@__PURE__*/
  enumDesc(file_ledger_network, 0);